Writeup
The night before the quiz, I practiced writing code for the movie examples that was provided for us. I believe that this allowed for me to have a good idea of what to expect and was how I approached this project. Much of the code is stemmed off of the lemonade stand project. I think that project laid a good foundation for this. I began in class with writing the kiosk class. As I was doing this in class, I made some mistakes with the user input, that actually made me write too many System.out.print lines. I realized these mistakes that I had made by the time I was writing the Printer class, and by then, it was time to turn it in. So, as I was thinking at home, I realized that with the use of a while loop and simple question, I can obtain all the user input I need. And so, I did just this. I didn't get to finish the printer class during class time, so I had a lot of thinking to do about how to take on this class. I understood that it was much like the tester class of lemonade stand and was able to identify what elements I need. I also needed to performs some math to decide how many spaces were needed and where were they needed. Although I feel I have made much progress, I still have some question. How can I ask the user how much of an item they want and implement it into my code? How to I call upon variables from another class? And, how do I use the String.length() function to calculate spaces more efficiently? I am hoping to tackle these questions as I continue to debug and work through this code.